Sprites map removed and replaced with maps for specific class types. Game works and hornet shoots bullets. Turret temporarily disappears when firing, though. Need fix for that.

+++ Group Game Engine/Ship.cpp
@@ -11,7 +11,7 @@ Ship::Ship()
 }
 
 
-Ship::Ship(float entID, int lvl, std::string entType, SDL_Texture* sprtsht, SDL_Rect spriteR, SDL_Rect hitBoxR, SDL_Renderer* ren)
+Ship::Ship(float entID, float lvl, std::string entType, SDL_Texture* sprtsht, SDL_Rect spriteR, SDL_Rect hitBoxR, SDL_Renderer* ren)
 {
 	id = entID;
 	layer = lvl;
